refactor(frontend): share showcase chat page
