+++ frontend/src/components/workspace/chats/use-thread-chat.ts
@@ -3,6 +3,7 @@
 import { useParams, usePathname, useSearchParams } from "next/navigation";
 import { useCallback, useEffect, useRef, useState } from "react";
 
+import { SHOWCASE_ROUTE_PREFIX } from "@/core/threads/static-demo";
 import { uuid } from "@/core/utils/uuid";
 
 export const THREAD_CHAT_RESET_EVENT = "deer-flow:thread-chat-reset";
@@ -120,7 +121,7 @@ export function useThreadChat() {
   }, []);
 
   const isMock =
-    actualPathname.startsWith("/showcase/") ||
+    actualPathname.startsWith(`${SHOWCASE_ROUTE_PREFIX}/`) ||
     searchParams.get("mock") === "true";
   return {
     threadId: isNewPath ? (newThreadIdRef.current ?? threadId) : threadId,
